+++ tests/core/parser/test_get_ignored_block_with_ignored_block.py
@@ -5,14 +5,15 @@ from assertpy import assert_that
 from printlinter import IgnoredBlock, get_ignored_block
 
 
-def test_get_ignored_block_with_ignored_block__ALL(
-    file_with_ignored_block_all,
+def test_get_ignored_block_with_ignored_block__ALL__re_enable(
+    file_with_ignored_block_all__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_all,
-            testing_files / "ignored_block/all.py",
+            file_with_ignored_block_all__re_enable,
+            testing_files / "ignored_block/re_enable/all.py",
+            18,
         )
     ).contains_only(
         *[
@@ -20,20 +21,21 @@ def test_get_ignored_block_with_ignored_block__ALL(
                 error_code="ALL",
                 line_from=10,
                 line_to=17,
-                from_file=testing_files / "ignored_block/all.py",
+                from_file=testing_files / "ignored_block/re_enable/all.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__ppl001(
-    file_with_ignored_block_print,
+def test_get_ignored_block_with_ignored_block__ppl001__re_enable(
+    file_with_ignored_block_print__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_print,
-            testing_files / "ignored_block/ppl001.py",
+            file_with_ignored_block_print__re_enable,
+            testing_files / "ignored_block/re_enable/ppl001.py",
+            9,
         )
     ).contains_only(
         *[
@@ -41,20 +43,21 @@ def test_get_ignored_block_with_ignored_block__ppl001(
                 error_code="PPL001",
                 line_from=5,
                 line_to=8,
-                from_file=testing_files / "ignored_block/ppl001.py",
+                from_file=testing_files / "ignored_block/re_enable/ppl001.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__ppl002(
-    file_with_ignored_block_pprint,
+def test_get_ignored_block_with_ignored_block__ppl002__re_enable(
+    file_with_ignored_block_pprint__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_pprint,
-            testing_files / "ignored_block/ppl002.py",
+            file_with_ignored_block_pprint__re_enable,
+            testing_files / "ignored_block/re_enable/ppl002.py",
+            13,
         )
     ).contains_only(
         *[
@@ -62,20 +65,21 @@ def test_get_ignored_block_with_ignored_block__ppl002(
                 error_code="PPL002",
                 line_from=9,
                 line_to=12,
-                from_file=testing_files / "ignored_block/ppl002.py",
+                from_file=testing_files / "ignored_block/re_enable/ppl002.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__ppl003(
-    file_with_ignored_block_stdout_write,
+def test_get_ignored_block_with_ignored_block__ppl003__re_enable(
+    file_with_ignored_block_stdout_write__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_stdout_write,
-            testing_files / "ignored_block/ppl003.py",
+            file_with_ignored_block_stdout_write__re_enable,
+            testing_files / "ignored_block/re_enable/ppl003.py",
+            13,
         )
     ).contains_only(
         *[
@@ -83,20 +87,21 @@ def test_get_ignored_block_with_ignored_block__ppl003(
                 error_code="PPL003",
                 line_from=9,
                 line_to=12,
-                from_file=testing_files / "ignored_block/ppl003.py",
+                from_file=testing_files / "ignored_block/re_enable/ppl003.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__ppl004(
-    file_with_ignored_block_stderr_write,
+def test_get_ignored_block_with_ignored_block__ppl004__re_enable(
+    file_with_ignored_block_stderr_write__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_stderr_write,
-            testing_files / "ignored_block/ppl004.py",
+            file_with_ignored_block_stderr_write__re_enable,
+            testing_files / "ignored_block/re_enable/ppl004.py",
+            13,
         )
     ).contains_only(
         *[
@@ -104,20 +109,21 @@ def test_get_ignored_block_with_ignored_block__ppl004(
                 error_code="PPL004",
                 line_from=9,
                 line_to=12,
-                from_file=testing_files / "ignored_block/ppl004.py",
+                from_file=testing_files / "ignored_block/re_enable/ppl004.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__ppl005(
-    file_with_ignored_block_stdout_writelines,
+def test_get_ignored_block_with_ignored_block__ppl005__re_enable(
+    file_with_ignored_block_stdout_writelines__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_stdout_writelines,
-            testing_files / "ignored_block/ppl005.py",
+            file_with_ignored_block_stdout_writelines__re_enable,
+            testing_files / "ignored_block/re_enable/ppl005.py",
+            13,
         )
     ).contains_only(
         *[
@@ -125,20 +131,21 @@ def test_get_ignored_block_with_ignored_block__ppl005(
                 error_code="PPL005",
                 line_from=9,
                 line_to=12,
-                from_file=testing_files / "ignored_block/ppl005.py",
+                from_file=testing_files / "ignored_block/re_enable/ppl005.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__ppl006(
-    file_with_ignored_block_stderr_writelines,
+def test_get_ignored_block_with_ignored_block__ppl006__re_enable(
+    file_with_ignored_block_stderr_writelines__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_stderr_writelines,
-            testing_files / "ignored_block/ppl006.py",
+            file_with_ignored_block_stderr_writelines__re_enable,
+            testing_files / "ignored_block/re_enable/ppl006.py",
+            13,
         )
     ).contains_only(
         *[
@@ -146,20 +153,21 @@ def test_get_ignored_block_with_ignored_block__ppl006(
                 error_code="PPL006",
                 line_from=9,
                 line_to=12,
-                from_file=testing_files / "ignored_block/ppl006.py",
+                from_file=testing_files / "ignored_block/re_enable/ppl006.py",
             )
         ]
     )
 
 
-def test_get_ignored_block_with_ignored_block__mix(
-    file_with_ignored_block_mix,
+def test_get_ignored_block_with_ignored_block__mix__re_enable(
+    file_with_ignored_block_mix__re_enable,
     testing_files,
 ):
     assert_that(
         get_ignored_block(
-            file_with_ignored_block_mix,
-            testing_files / "ignored_block/mix.py",
+            file_with_ignored_block_mix__re_enable,
+            testing_files / "ignored_block/re_enable/mix.py",
+            17,
         )
     ).contains_only(
         *[
@@ -167,13 +175,195 @@ def test_get_ignored_block_with_ignored_block__mix(
                 error_code="PPL002",
                 line_from=9,
                 line_to=16,
-                from_file=testing_files / "ignored_block/mix.py",
+                from_file=testing_files / "ignored_block/re_enable/mix.py",
             ),
             IgnoredBlock(
                 error_code="PPL001",
                 line_from=12,
                 line_to=15,
-                from_file=testing_files / "ignored_block/mix.py",
+                from_file=testing_files / "ignored_block/re_enable/mix.py",
+            ),
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ALL__no_re_enable(
+    file_with_ignored_block_all__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_all__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/all.py",
+            17,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="ALL",
+                line_from=10,
+                line_to=17,
+                from_file=testing_files / "ignored_block/no_re_enable/all.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ppl001__no_re_enable(
+    file_with_ignored_block_print__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_print__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/ppl001.py",
+            8,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL001",
+                line_from=5,
+                line_to=8,
+                from_file=testing_files / "ignored_block/no_re_enable/ppl001.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ppl002__no_re_enable(
+    file_with_ignored_block_pprint__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_pprint__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/ppl002.py",
+            12,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL002",
+                line_from=9,
+                line_to=12,
+                from_file=testing_files / "ignored_block/no_re_enable/ppl002.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ppl003__no_re_enable(
+    file_with_ignored_block_stdout_write__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_stdout_write__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/ppl003.py",
+            12,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL003",
+                line_from=9,
+                line_to=12,
+                from_file=testing_files / "ignored_block/no_re_enable/ppl003.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ppl004__no_re_enable(
+    file_with_ignored_block_stderr_write__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_stderr_write__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/ppl004.py",
+            12,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL004",
+                line_from=9,
+                line_to=12,
+                from_file=testing_files / "ignored_block/no_re_enable/ppl004.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ppl005__no_re_enable(
+    file_with_ignored_block_stdout_writelines__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_stdout_writelines__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/ppl005.py",
+            12,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL005",
+                line_from=9,
+                line_to=12,
+                from_file=testing_files / "ignored_block/no_re_enable/ppl005.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__ppl006__no_re_enable(
+    file_with_ignored_block_stderr_writelines__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_stderr_writelines__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/ppl006.py",
+            12,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL006",
+                line_from=9,
+                line_to=12,
+                from_file=testing_files / "ignored_block/no_re_enable/ppl006.py",
+            )
+        ]
+    )
+
+
+def test_get_ignored_block_with_ignored_block__mix__no_re_enable(
+    file_with_ignored_block_mix__no_re_enable,
+    testing_files,
+):
+    assert_that(
+        get_ignored_block(
+            file_with_ignored_block_mix__no_re_enable,
+            testing_files / "ignored_block/no_re_enable/mix.py",
+            15,
+        )
+    ).contains_only(
+        *[
+            IgnoredBlock(
+                error_code="PPL002",
+                line_from=9,
+                line_to=15,
+                from_file=testing_files / "ignored_block/no_re_enable/mix.py",
+            ),
+            IgnoredBlock(
+                error_code="PPL001",
+                line_from=12,
+                line_to=15,
+                from_file=testing_files / "ignored_block/no_re_enable/mix.py",
             ),
         ]
     )