+++ tests/testing_files/ignored_block/no_re_enable/ppl006.py
@@ -1,0 +1,11 @@
+# Standard imports
+from sys import stderr
+
+
+def toto():
+    return 1 + 2
+
+
+# <py-printlinter disable PPL006>
+stderr.writelines(toto())
+stderr.writelines(toto() + toto())